Finished Stealth Attack
The stealth attack is fully finished and fixed a bug with the stealth attack that caused the player to go throw the floor. Also made a simple main menu (options is the only thing that doesnt work) and made the double jump play the jump animation on the secound jump.