Made options menu, Save and load system, Equip able weapons and boots.
Made the options menu only has texture and screen res at the moment (easy to add more stuff) made the save and load feature my pressing o to save and j to load, only saves player position, but easy to save everything else. Made item pick up system with a sword and shoes. The shoes need to be fixed so that they fit both feet. When the player has boots equiped the player can double jump. Need to change it so that the player changes attack animations and hitboxes when picking up weapon