Combo Attack System
Made a combo attack system using the did it hit plugin, the system if very modular for new weapons and animations. Idk if the plugin auto installs but if it doesnt find the unreal engine source file on your computer and find the 5.1 and on the plugins folder add the folder thats called marketplace which is in the git plugins folder. Also some minor bug changes.